1. class UserDaoImpl.void update(int id, User updatedUser) реализуй метод без id в параметре - вырежи id вплоть до контроллера используй метод merge 2. User class используй в качестве id тип Long

+++ src/main/java/web/model/User.java
@@ -8,7 +8,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
@@ -21,18 +21,18 @@ public class User {
 
     public User(){}
 
-    public User(int id, String name, String lastName, String email) {
+    public User(Long id, String name, String lastName, String email) {
         this.id = id;
         this.name = name;
         this.lastName = lastName;
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
